"unpurge" packages marked for purge in pkg.installed states
See https://github.com/saltstack/salt/issues/10719.

+++ salt/states/pkg.py
@@ -40,6 +40,7 @@ import re
 
 # Import salt libs
 import salt.utils
+from salt._compat import string_types
 from salt.exceptions import CommandExecutionError, MinionError
 from salt.modules.pkg_resource import _repack_pkgs
 
@@ -95,6 +96,19 @@ def _fulfills_version_spec(versions, oper, desired_version):
     return False
 
 
+def _find_unpurge_targets(desired):
+    '''
+    Find packages which are marked to be purged but can't yet be removed
+    because they are dependencies for other installed packages. These are the
+    packages which will need to be 'unpurged' because they are part of
+    pkg.installed states. This really just applies to Debian-based Linuxes.
+    '''
+    return [
+        x for x in desired
+        if x in __salt__['pkg.list_pkgs'](purge_desired=True)
+    ]
+
+
 def _find_install_targets(name=None,
                           version=None,
                           pkgs=None,
@@ -126,7 +140,7 @@ def _find_install_targets(name=None,
                     'comment': 'Invalidly formatted {0!r} parameter. See '
                                'minion log.'.format('pkgs' if pkgs
                                                     else 'sources')}
-
+        to_unpurge = _find_unpurge_targets(desired)
     else:
         if salt.utils.is_windows():
             pkginfo = _get_package_info(name)
@@ -139,23 +153,26 @@ def _find_install_targets(name=None,
             if version is None:
                 version = _get_latest_pkg_version(pkginfo)
         desired = {name: version}
+        to_unpurge = _find_unpurge_targets(desired)
 
         cver = cur_pkgs.get(name, [])
-        if version and version in cver:
-            # The package is installed and is the correct version
-            return {'name': name,
-                    'changes': {},
-                    'result': True,
-                    'comment': ('Version {0} of package {1!r} is already '
-                                'installed').format(version, name)}
+        if name not in to_unpurge:
+            if version and version in cver:
+                # The package is installed and is the correct version
+                return {'name': name,
+                        'changes': {},
+                        'result': True,
+                        'comment': 'Version {0} of package {1!r} is already '
+                                   'installed'.format(version, name)}
 
-        # if cver is not an empty string, the package is already installed
-        elif cver and version is None:
-            # The package is installed
-            return {'name': name,
-                    'changes': {},
-                    'result': True,
-                    'comment': 'Package {0} is already installed'.format(name)}
+            # if cver is not an empty string, the package is already installed
+            elif cver and version is None:
+                # The package is installed
+                return {'name': name,
+                        'changes': {},
+                        'result': True,
+                        'comment': 'Package {0} is already '
+                                   'installed'.format(name)}
 
     version_spec = False
     # Find out which packages will be targeted in the call to pkg.install
@@ -228,16 +245,18 @@ def _find_install_targets(name=None,
                     'result': False,
                     'comment': ' '.join(problems)}
 
-    if not targets:
+    if not any((targets, to_unpurge)):
         # All specified packages are installed
-        msg = 'All specified packages are already installed{0}.'.format(
-            ' and are at the desired version' if version_spec else '')
+        msg = (
+            'All specified packages are already installed{0}.'
+            .format(' and are at the desired version' if version_spec else '')
+        )
         return {'name': name,
                 'changes': {},
                 'result': True,
                 'comment': msg}
 
-    return desired, targets
+    return desired, targets, to_unpurge
 
 
 def _verify_install(desired, new_pkgs):
@@ -436,7 +455,7 @@ def installed(
         or (os.path.isfile(rtag) and refresh is not False)
     )
 
-    if not isinstance(version, basestring) and version is not None:
+    if not isinstance(version, string_types) and version is not None:
         version = str(version)
 
     result = _find_install_targets(name, version, pkgs, sources,
@@ -444,10 +463,16 @@ def installed(
                                    skip_suggestions=skip_suggestions,
                                    **kwargs)
     try:
-        desired, targets = result
+        desired, targets, to_unpurge = result
     except ValueError:
         # _find_install_targets() found no targets or encountered an error
         return result
+
+    if to_unpurge and 'lowpkg.unpurge' not in __salt__:
+        return {'name': name,
+                'changes': {},
+                'result': False,
+                'comment': 'lowpkg.unpurge not implemented'}
 
     # Remove any targets that are already installed, to avoid upgrading them
     if pkgs:
@@ -455,6 +480,7 @@ def installed(
     elif sources:
         sources = [x for x in sources if x.keys()[0] in targets]
 
+    comment = []
     if __opts__['test']:
         if targets:
             if sources:
@@ -462,45 +488,50 @@ def installed(
             else:
                 summary = ', '.join([_get_desired_pkg(x, targets)
                                      for x in targets])
-            comment = 'The following packages are set to be ' \
-                      'installed/updated: {0}.'.format(summary)
-        else:
-            comment = ''
+            comment.append('The following packages are set to be '
+                           'installed/updated: {0}.'.format(summary))
+        if to_unpurge:
+            comment.append(
+                'The following packages will have their selection status '
+                'changed from \'purge\' to \'install\': {0}.'
+                .format(', '.join(to_unpurge))
+            )
         return {'name': name,
                 'changes': {},
                 'result': None,
-                'comment': comment}
+                'comment': ' '.join(comment)}
 
-    comment = []
-    try:
-        pkg_ret = __salt__['pkg.install'](name,
-                                          refresh=refresh,
-                                          version=version,
-                                          fromrepo=fromrepo,
-                                          skip_verify=skip_verify,
-                                          pkgs=pkgs,
-                                          sources=sources,
-                                          **kwargs)
+    changes = {'installed': {}}
+    if targets:
+        try:
+            pkg_ret = __salt__['pkg.install'](name,
+                                            refresh=refresh,
+                                            version=version,
+                                            fromrepo=fromrepo,
+                                            skip_verify=skip_verify,
+                                            pkgs=pkgs,
+                                            sources=sources,
+                                            **kwargs)
 
-        if os.path.isfile(rtag) and refresh:
-            os.remove(rtag)
-    except CommandExecutionError as exc:
-        return {'name': name,
-                'changes': {},
-                'result': False,
-                'comment': 'An error was encountered while installing '
-                           'package(s): {0}'.format(exc)}
+            if os.path.isfile(rtag) and refresh:
+                os.remove(rtag)
+        except CommandExecutionError as exc:
+            return {'name': name,
+                    'changes': {},
+                    'result': False,
+                    'comment': 'An error was encountered while installing '
+                            'package(s): {0}'.format(exc)}
 
-    if isinstance(pkg_ret, dict):
-        changes = pkg_ret
-    elif isinstance(pkg_ret, basestring):
-        changes = {}
-        comment.append(pkg_ret)
-    else:
-        changes = {}
+        if isinstance(pkg_ret, dict):
+            changes['installed'].update(pkg_ret)
+        elif isinstance(pkg_ret, string_types):
+            comment.append(pkg_ret)
+
+    if to_unpurge:
+        changes['purge_desired'] = __salt__['lowpkg.unpurge'](*to_unpurge)
 
     if sources:
-        modified = [x for x in changes.keys() if x in targets]
+        modified = [x for x in changes['installed'].keys() if x in targets]
         not_modified = [x for x in desired if x not in targets]
         failed = [x for x in targets if x not in modified]
     else:
@@ -512,6 +543,11 @@ def installed(
             )
         modified = [x for x in ok if x in targets]
         not_modified = [x for x in ok if x not in targets]
+
+    # If there was nothing unpurged, just set the changes dict to the contents
+    # of changes['installed'].
+    if not changes.get('purge_desired'):
+        changes = changes['installed']
 
     if modified:
         if sources:
@@ -645,9 +681,9 @@ def latest(
         os.remove(rtag)
 
     # Repack the cur/avail data if only a single package is being checked
-    if isinstance(cur, basestring):
+    if isinstance(cur, string_types):
         cur = {desired_pkgs[0]: cur}
-    if isinstance(avail, basestring):
+    if isinstance(avail, string_types):
         avail = {desired_pkgs[0]: avail}
 
     targets = {}
